+++ wp-content/plugins/github-updater/src/GitHub_Updater/Language_Pack.php
@@ -38,19 +38,23 @@ class Language_Pack extends Base {
 	 * @param Language_Pack_API $api  Language_Pack_API object.
 	 */
 	public function __construct( $repo, Language_Pack_API $api ) {
-		if ( empty( $repo->languages ) ) {
+		parent::__construct();
+		if ( null === $repo->languages ) {
 			return;
 		}
 
 		$this->repo     = $repo;
 		$this->repo_api = $api;
-		$this->run();
 	}
 
 	/**
 	 * Do the Language Pack integration.
 	 */
-	protected function run() {
+	public function run() {
+		if ( null === $this->repo ) {
+			return false;
+		}
+
 		$headers = $this->parse_header_uri( $this->repo->languages );
 		$this->repo_api->get_language_pack( $headers );
 
@@ -75,11 +79,11 @@ class Language_Pack extends Base {
 		}
 
 		if ( 'pre_set_site_transient_update_plugins' === current_filter() ) {
-			$repos        = Plugin::instance()->get_plugin_configs();
+			$repos        = Singleton::get_instance( 'Plugin' )->get_plugin_configs();
 			$translations = wp_get_installed_translations( 'plugins' );
 		}
 		if ( 'pre_set_site_transient_update_themes' === current_filter() ) {
-			$repos        = Theme::instance()->get_theme_configs();
+			$repos        = Singleton::get_instance( 'Theme' )->get_theme_configs();
 			$translations = wp_get_installed_translations( 'themes' );
 		}